% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codings.R
\name{bio_code_primary_care}
\alias{bio_code_primary_care}
\title{Reads UKB reference data: primary care maps and lookups}
\usage{
bio_code_primary_care(project_dir, lkp = NULL)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{lkp}{A character vector naming the required table. To see all
available primary care coding tables, set \code{lkp = NULL} (default).}
}
\value{
If \code{lkp = NULL} (default) a data frame of available primary
care lookup tables. If a value from the \code{table} column is supplied
to \code{lkp}, a data frame of the named primary care lookup table:
}
\description{
Reads UKB reference data: primary care maps and lookups
}
