% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{bio_death}
\alias{bio_death}
\title{Reads death records}
\usage{
bio_death(project_dir, record = "death", death_dir = "raw/")
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{record}{A string specifying which death data are required:
"death" (default) includes date of death, "cause" includes ICD-10
cause of death.}

\item{death_dir}{Path to the enclosing directory of the death data.}
}
\value{
A dataframe of including either the date of death, or cause
of death.
}
\description{
Reads death records
}
\seealso{
\href{https://biobank.ctsu.ox.ac.uk/crystal/label.cgi?id=100093}{Category 100093},
\href{https://biobank.ctsu.ox.ac.uk/crystal/refer.cgi?id=134993}{Resource 134993}
}
