% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_read.R
\name{bio_field}
\alias{bio_field}
\title{Read the UKB field codes}
\usage{
bio_field(project_dir)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project}
}
\value{
A dataframe with columns: ukb_type, r_type, path
}
\description{
The field code lookup table associates with each field id, a desciptive name
(which includes the field id, array, index), ukb_type, r_type, and the path
to the csv file containing the field data.
}
