% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_read.R
\name{bio_gen}
\alias{bio_gen}
\title{Prints paths to genetic data}
\usage{
bio_gen(project_dir)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}
}
\description{
Prints paths to genetic data
}
