% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{bio_gen_ancestry}
\alias{bio_gen_ancestry}
\title{Assigns 1000 Genomes super populations}
\usage{
bio_gen_ancestry(project_dir)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}
}
\value{
A dataframe of project-specific pseudo-IDs and 1000 Genomes super
population
}
\description{
Assigns 1000 genomes super population after ancestry-specific QC. For
QC details, see Ollie Pain's
\href{https://opain.github.io/UKB-GenoPrep/quality_control.html}{Ancestry Specific Quality Control}
documentation.
}
