% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{bio_gen_related}
\alias{bio_gen_related}
\title{Read the project-specific relatedness file}
\usage{
bio_gen_related(project_dir)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}
}
\value{
A dataframe of relatedness with project-specific pseudo-IDS.
}
\description{
Read the project-specific relatedness file
}
