% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{bio_gen_write_plink_input}
\alias{bio_gen_write_plink_input}
\title{Writes a two-column dataframe of IDs for PLINK input}
\usage{
bio_gen_write_plink_input(data, out)
}
\arguments{
\item{data}{Either a vector of sample IDs, or a dataframe with
sample IDs in column 1.}

\item{out}{Full file path to write PLINK input sample list to.}
}
\description{
Writes a two-column dataframe of IDs for PLINK input
}
