% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_read.R
\name{bio_gp}
\alias{bio_gp}
\title{Reads the primary care data}
\usage{
bio_gp(project_dir, record)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{record}{A string specifying which primary care records are required: "clinical", "registrations", "scripts".}
}
\value{
A dataframe. \strong{Note}. clinical data has 123,669,371 rows and 8 columns; registrations data has 361,841 rows and 4 columns; scripts data has 57,709,810 rows and 8 columns.
}
\description{
Detailed patient level diagnoses, prescriptions, etc. Only available if these data have been requested for the particular project you have access to.
}
