% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_read.R
\name{bio_phen}
\alias{bio_phen}
\title{Reads and writes phenotype data for a subset of fields}
\usage{
bio_phen(
  project_dir,
  field_subset_file,
  pheno_dir = "phenotypes",
  out = "ukb_phenotype_subset"
)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{field_subset_file}{A path to a one-per-line text file of
fields (no header).}

\item{pheno_dir}{Path to the enclosing directory of the phenotype
data.}

\item{out}{Name of phenotype subset file. Default
"ukb_phenotype_subset", writes ukb_phenotype_subset.rds to the
current directory.}
}
\description{
Reads supplied fields from UKB project data and writes a serialized
dataframe to a .rds file.
}
\details{
Read the serialized dataframe with
readRDS("<name_of_phenotype_subset_file>.rds")
}
