% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{bio_record_map}
\alias{bio_record_map}
\title{Applies a function to each record-level disk.frame}
\usage{
bio_record_map(project_dir, func, records = "all")
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{func}{A function to apply to each element in the \code{records}
vector, e.g. names, head.}

\item{records}{A character vector of record-level data to apply
summary function to. Default value \code{"all"} applies \code{func} to all
available record-level data.}
}
\description{
Applies a function to each record-level disk.frame
}
\seealso{
\link{bio_record}
}
