% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{bio_return}
\alias{bio_return}
\title{Reads returned data}
\usage{
bio_return(project_dir, return = NULL)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{return}{A integer indicating which UKB return to read, e.g.,
\code{3388} for PGxPOP returned allele and phenotype calls.}
}
\value{
For return \code{3388} a data frame of
\href{https://github.com/PharmGKB/PGxPOP}{PGxPOP}
called diplotypes and phenotypes, for both the imputed and integrated
(imputed plus exome) data. See \href{https://pubmed.ncbi.nlm.nih.gov/33237584/}{McInnes et al (2020) Pharmacogenetics at scale: An analysis of the UK Biobank}
}
\description{
Reads returned data
}
