% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_dmd_antidep}
\alias{drug_dmd_antidep}
\title{Curated dm+d antidepressant names}
\format{
An object of class \code{data.frame} with 137 rows and 1 columns.
}
\usage{
drug_dmd_antidep
}
\description{
Antidepressant dm+d names extracted by \href{https://pubmed.ncbi.nlm.nih.gov/33753889/}{Fabbri et al. (2021) Genetic and clinical characteristics of treatment-resistant depression using primary care records in two UK cohorts}
\describe{
\item{dmd_name}{Category}
}
}
\keyword{datasets}
