% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_gwas}
\alias{drug_gwas}
\title{Active ingredients and ATC code of UKB self-reported medications}
\format{
A data frame with 1752 obs. of  4 variables:
\describe{
\item{category}{Category}
\item{ukb_code}{Coding reported by UK Biobank.}
\item{atc_code}{Medication ATC code}
\item{name}{Drug name}
}
}
\usage{
drug_gwas
}
\description{
A dataset containing
\href{10.1038/s41467-019-09572-5}{Wu et al. (2019) Nat Commun},
Supplementary Data 1: Active ingredients and ATC code of medications
in UK Biobank. This study was a series of "GWASs of self-reported
medication use from 23 medication categories".
}
\keyword{datasets}
