% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_pharmgkb}
\alias{drug_pharmgkb}
\title{PharmGKB curated drug list}
\format{
A data frame with 3493 obs. of  24 variables:
\describe{
\item{pharmgkb_accession_id}{Identifier assigned to this chemical by PharmGKB}
\item{name}{Name PharmGKB uses for this drug}
\item{generic_names}{Known generic names for this drug, comma-separated and "-enclosed}
\item{trade_names}{Known trade names for this drug, comma-separated and "-enclosed}
\item{brand_mixtures}{Known brand mixtures this drug is in, comma-separated and "-enclosed}
\item{type}{Categories PharmGKB has assigned to this drug, can be more than one, possible values: Drug, Metabolite,
Ion, Drug Class, Biological Intermediate, Small Molecule, Prodrug}
\item{cross_references}{References to other resources in the form "resource:id", comma-separated}
\item{smiles}{The SMILES structure for this drug}
\item{inchi}{The InCHI key for this drug}
\item{dosing_guideline}{"Yes" if PharmGKB has annotated a guideline with this drug, "No" otherwise}
\item{external_vocabulary}{Term for this drug in another vocabulary in the form "vocabulary:id", comma-separated}
}

Multiple "Type" values can be assigned to a given entry
since the same substance can be used in different contexts. The
"Type" values used for Drugs & Chemicals are as follows:
\itemize{
\item \strong{Drug} A chemical substance used in the treatment, cure,prevention, or diagnosis of disease.
\item \strong{Metabolite} Any intermediate or product resulting from metabolism.
\item \strong{Ion} An atomic or molecular particle having a net electric charge.
\item \strong{Drug Class} A drug class is a group of medications that may work in the same way, have a similar chemical structure, or are used to treat the same health condition.
\item \strong{Biological Intermediate} An endogenous small molucule or ion.
\item \strong{Small Molecule} An electrically neutral entity consisting of more than one atom.
\item \strong{Prodrug} A compound that must undergo chemical conversion by metabolic processes before becoming the pharmacologically active drug for which it is a prodrug.
}
}
\source{
\url{https://www.pharmgkb.org/downloads}
}
\usage{
drug_pharmgkb
}
\description{
A dataset containing all Chemicals with the Type of "Drug" in the PharmGKB
knowledgebase. Not all of these drugs have been involved in PharmGKB
annotations.
}
\keyword{datasets}
