% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_ukb}
\alias{drug_ukb}
\title{UKB primary care drug maps and lookups}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\source{
\url{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}
}
\usage{
drug_ukb
}
\description{
A nested tibble of the UKB primary care coding lookups and maps
for drug coding information. The tables have been read into
data frames from ll_lkps_map_v3.xlsx included in primarycare_codings.zip,
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{Resource 592}.
}
\details{
\describe{
\item{df}{Data frame from coding map}
\item{name}{Sheet name from all_lkps_map_v3.xlsx included in primarycare_codings.zip}
\item{description}{Description of coding map version}
}
}
\keyword{datasets}
