% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbkings.R
\docType{package}
\name{ukbkings}
\alias{ukbkings}
\title{ukbkings: a King's College London interface to UKB project data}
\description{
The ukbkings package provides a few basic tools to access UKB project data on
Rosalind HPC, for UKB-approved KCL researchers and collaborators. Access to a
particular project is restricted to named KCL researchers and collaborators,
approved by the UKB, for a particular study.
}
\section{ukbkings functions}{

bio_field
bio_code
bio_data
}

