% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{bio_code}
\alias{bio_code}
\title{Reads the UKB showcase codings for categorical variables}
\usage{
bio_code(code_dir = "/scratch/datasets/ukbiobank/resources")
}
\arguments{
\item{code_dir}{Path to the enclosing directory of the
Codings_Showcase.csv.}
}
\value{
A dataframe with header Coding, Value, Meaning
}
\description{
Reads the UKB showcase codings for categorical variables
}
