% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{bio_gp}
\alias{bio_gp}
\title{Reads the primary care data}
\usage{
bio_gp(project_dir, record, gp_dir = "raw")
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{record}{A string specifying which primary care records are
required: "clinical", "registrations", "scripts".}

\item{gp_dir}{Path to the enclosing directory of the primary care
data.}
}
\value{
A dataframe. \strong{Note}. clinical data has 123,669,371
rows and 8 columns; registrations data has 361,841 rows and 4
columns; scripts data has 57,709,810 rows and 8 columns.
}
\description{
#' @description
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}
}
\details{
This function has been deprecated in favour of a single function to
read all record-level data. Use \code{\link[=bio_record]{bio_record()}} to retrieve data.
Summarize/ inspect record-level data with \code{\link[=bio_record_map]{bio_record_map()}}.

Detailed patient level diagnoses, prescriptions, etc. Only available
if these data have been requested for the particular project you
have access to.
}
\seealso{
\href{http://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=3001}{Category 3001},
\href{http://biobank.ndph.ox.ac.uk/showcase/refer.cgi?id=591}{Resource 591},
\href{http://biobank.ndph.ox.ac.uk/showcase/refer.cgi?id=592}{Resource 592}
}
\keyword{internal}
