% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{bio_record}
\alias{bio_record}
\title{Reads record-level data from on-disk disk.frames}
\usage{
bio_record(project_dir, record)
}
\arguments{
\item{project_dir}{Path to the enclosing directory of a UKB project.}

\item{record}{A string specifying which record-level data are required:
\bold{COVID-19} data ("covid19_emis_gp_clinical", "covid19_result",
"covid19_emis_gp_scripts", "covid19_tpp_gp_clinical", "covid19_misc",
"covid19_tpp_gp_scripts"),
\bold{primary care} data ("gp_clinical", "gp_scripts", "gp_registrations"),
\bold{hospital episode statistics} data ("hesin", "hesin_critical",
"hesin_delivery", "hesin_diag", "hesin_maternity", "hesin_oper",
"hesin_psych"),
\bold{death} data ("death", "death_cause")}
}
\value{
A disk.frame of the specified record.
}
\description{
Reads record-level data from on-disk disk.frames
}
\details{
...
}
