% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_read.R
\name{bio_rename}
\alias{bio_rename}
\title{Updates column names}
\usage{
bio_rename(data, field_finder)
}
\arguments{
\item{data}{A dataframe with UKB fields as column names. See \code{\link{bio_phen}}.}

\item{field_finder}{A dataframe including a column of UKB fields and a column of descriptive names. See \code{\link{bio_field}}.}
}
\value{
A dataframe with UKB field column names replaced with descriptive column names.
importFrom dplyr filter rename
}
\description{
Renames a dataframe with UKB fields as column names with descriptive names.
}
\seealso{
\code{\link{bio_phen}}, \code{\link{bio_field}}
}
